requirements.txt
torch
transformers>=4.37.0
pillow
accelerate
tqdm
huggingface_hub